refactor(pfb): extract section snippets into a useSnippets composable
